+++ AmazonChimeSDK/AmazonChimeSDK/ingestion/IngestionEvent.swift
@@ -13,16 +13,24 @@ import Foundation
     /// Event Client Type associated with this event
     public let type: String
     /// metadata that could be overriden
-    public let metadata: IngestionMetadata
+    public let metadata: [String: AnyCodable?]
     /// Payload associated with this ingestion event.
     public let payloads: [IngestionPayload]
     /// Version of payload. Different event format could give different version.
     public let version: Int
-    public convenience init(type: String, metadata: IngestionMetadata, payloads: [IngestionPayload]) {
-        self.init(type: type, metadata: metadata, payloads: payloads, version: 1)
+    public convenience init(type: String,
+                            metadata: [String: AnyCodable?],
+                            payloads: [IngestionPayload]) {
+        self.init(type: type,
+                  metadata: metadata,
+                  payloads: payloads,
+                  version: 1)
     }
 
-    public init(type: String, metadata: IngestionMetadata, payloads: [IngestionPayload], version: Int) {
+    public init(type: String,
+                metadata: [String: AnyCodable?],
+                payloads: [IngestionPayload],
+                version: Int) {
         self.type = type
         self.metadata = metadata
         self.payloads = payloads