+++ AmazonChimeSDK/AmazonChimeSDK/ingestion/IngestionRecord.swift
@@ -11,11 +11,11 @@ import Foundation
 /// `IngestionRecord` is the format of data that will be consumed on the ingestion server.
 @objcMembers public class IngestionRecord: NSObject, Codable {
     /// Metadata associated with the event. This includes deviceName, OSVersion, and etc.
-    public let metadata: IngestionMetadata
+    public let metadata: [String: AnyCodable?]
     /// List of `IngestionEvent`
     public let events: [IngestionEvent]
 
-    public init(metadata: IngestionMetadata, events: [IngestionEvent]) {
+    public init(metadata: [String: AnyCodable?], events: [IngestionEvent]) {
         self.events = events
         self.metadata = metadata
     }